precision landing: pequenos ajustes essenciais

+++ mirela_sdk/mirela_sdk/control/mavros/mavros_api.py
@@ -434,7 +434,7 @@ class MavDrone(Drone):
         
         Parameters
         ----------
-        rtl_alt: float (m)
+        rtl_alt: int (m)
 
         precisionland: bool (True or False)
         """
